Fetch stocks concurrently from one data provider
* Create fetcher functions for all data providers using Selenium, where multiple function calls can work on the same collections
* Add and make use of Fetch API parameter `concurrency` and environment variable `SELENIUM_MAX_CONCURRENCY`
* Put WebDriver-related functions in separate `utils` file
