+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,20 +1,19 @@
 ---
 name: Feature request
 about: Suggest an idea for this project
-title: ''
+title: ""
 labels: feature
 assignees: marvinruder
-
 ---
 
 **Is your feature request related to a problem? Please describe.**
-A clear and concise description of what the problem is. Ex. I'm always frustrated when [...]
+A clear and concise description of what the problem is. E.g. “I’m always frustrated when […]”
 
-**Describe the solution you'd like**
+**Describe the solution you’d like**
 A clear and concise description of what you want to happen.
 
-**Describe alternatives you've considered**
-A clear and concise description of any alternative solutions or features you've considered.
+**Describe alternatives you’ve considered**
+A clear and concise description of any alternative solutions or features you’ve considered.
 
 **Additional context**
 Add any other context or screenshots about the feature request here.
