+++ packages/rating-tracker-backend/src/controllers/FetchController.ts
@@ -2,12 +2,9 @@
 /* istanbul ignore file -- @preserve */
 import { formatDistance } from "date-fns";
 import { Request, Response } from "express";
-import { Builder, By, Capabilities, until, WebDriver } from "selenium-webdriver";
-import { Options } from "selenium-webdriver/chrome.js";
 import APIError from "../utils/apiError.js";
 import chalk from "chalk";
 import {
-  Currency,
   fetchMarketScreenerEndpointPath,
   fetchMorningstarEndpointPath,
   fetchMSCIEndpointPath,
@@ -15,17 +12,8 @@ import {
   fetchSPEndpointPath,
   fetchSustainalyticsEndpointPath,
   GENERAL_ACCESS,
-  Industry,
-  isCurrency,
-  isIndustry,
-  isMSCIESGRating,
-  isSize,
-  isStyle,
-  MSCIESGRating,
   Resource,
-  Size,
   Stock,
-  Style,
   WRITE_STOCKS_ACCESS,
 } from "rating-tracker-commons";
 import { readAllStocks, readStock, updateStock } from "../db/tables/stockTable.js";
@@ -35,102 +23,52 @@ import { createResource, readResource } from "../redis/repositories/resourceRepo
 import axios from "axios";
 import dotenv from "dotenv";
 import Router from "../utils/router.js";
+import { SIGNAL_PREFIX_ERROR } from "../signal/signal.js";
+import morningstarFetcher from "../fetchers/morningstarFetcher.js";
+import marketScreenerFetcher from "../fetchers/marketScreenerFetcher.js";
+import msciFetcher from "../fetchers/msciFetcher.js";
+import refinitivFetcher from "../fetchers/refinitivFetcher.js";
+import spFetcher from "../fetchers/spFetcher.js";
 
 dotenv.config();
-
-const SIGNAL_PREFIX_ERROR = "⚠️ " as const;
-const SIGNAL_PREFIX_INFO = "ℹ️ " as const;
-
-const XPATH_INDUSTRY = "//*/div[@id='CompanyProfile']/div/h3[contains(text(), 'Industry')]/.." as const;
-const XPATH_SIZE_STYLE = "//*/div[@id='CompanyProfile']/div/h3[contains(text(), 'Stock Style')]/.." as const;
-const XPATH_STAR_RATING = "//*/img[@class='starsImg']" as const;
-const XPATH_MORNINGSTAR_FAIR_VALUE = "//*/datapoint[@id='FairValueEstimate']" as const;
-const XPATH_DESCRIPTION = "//*/div[@id='CompanyProfile']/div[1][not(.//h3)]" as const;
-
-const XPATH_CONSENSUS_DIV = "//*/div[@class='tabTitleLeftWhite']/b[contains(text(), 'Consensus')]/../../../.." as const;
-const XPATH_CONSENSUS_NOTE = "//div/table/tbody/tr/td/div/div[starts-with(@title, 'Note : ')]" as const;
-const XPATH_ANALYST_COUNT =
-  "//div/table/tbody/tr/td[contains(text(), 'Number of Analysts')]/following-sibling::td" as const;
-const XPATH_SPREAD_AVERAGE_TARGET =
-  "//div/table/tbody/tr/td[contains(text(), 'Spread / Average Target')]/following-sibling::td" as const;
-
-const XPATH_SP_PANEL = "//*/div[@class='panel-set__first-column']/h1[@id='company-name']/.." as const;
 
 const URL_SUSTAINALYTICS = "https://www.sustainalytics.com/sustapi/companyratings/getcompanyratings" as const;
 
 /**
- * Creates and returns a new WebDriver instance.
- *
- * @param {boolean} headless whether to run the browser in headless mode
- * @returns {Promise<WebDriver>} a Promise that resolves to a WebDriver instance
- * @throws an {@link APIError} if the WebDriver cannot be created
+ * A shared object holding lists of stocks that multiple fetchers work on.
  */
-const getDriver = async (headless?: boolean) => {
-  const url = process.env.SELENIUM_URL;
-  const options = new Options().addArguments("window-size=1080x3840"); // convenient for screenshots
-  headless && options.headless(); // In headless mode, the browser window is not shown.
-
-  return await new Builder()
-    .usingServer(url)
-    .withCapabilities(
-      new Capabilities()
-        // Use Chrome as the browser.
-        .setBrowserName("chrome")
-        // Do not wait for all resources to load. This speeds up the page load.
-        .setPageLoadStrategy("eager")
-    )
-    .setChromeOptions(options)
-    .build()
-    .then((driver) => driver)
-    .catch((e) => {
-      throw new APIError(502, `Unable to connect to Selenium WebDriver: ${e.message}`);
-    });
+export type FetcherWorkspace<T> = {
+  queued: T[];
+  successful: T[];
+  failed: T[];
 };
 
 /**
- * Shuts down the given WebDriver instance gracefully, deallocating all associated resources.
+ * Determines the allowed number of fetchers that can work concurrently on fetching a list of stocks.
  *
- * @param {WebDriver} driver the WebDriver instance to shut down
- * @returns {Promise<void>} a Promise that resolves when the WebDriver has been shut down
- * @throws an {@link APIError} if the WebDriver cannot be shut down gracefully
+ * @param {Request} req Request object
+ * @returns {number} The number of fetchers to use.
  */
-const quitDriver = async (driver: WebDriver): Promise<void> => {
-  try {
-    await driver.quit();
-  } catch (e) {
-    logger.warn(PREFIX_SELENIUM + chalk.yellowBright(`Unable to shut down Selenium WebDriver gracefully: ${e}`));
-  }
-};
-
-/**
- * Creates a screenshot of the current page and stores it in Redis.
- *
- * @param {WebDriver} driver the WebDriver instance in use
- * @param {Stock} stock the affected stock
- * @param {string} dataProvider the name of the data provider
- * @returns {Promise<string>} a Promise that resolves to a string holding a general informational message and a URL to
- * the screenshot
- */
-const takeScreenshot = async (driver: WebDriver, stock: Stock, dataProvider: string): Promise<string> => {
-  const screenshotID = `error-${dataProvider}-${stock.ticker}-${new Date().getTime().toString()}.png`;
-  try {
-    const screenshot = await driver.takeScreenshot();
-    // deepcode ignore Ssrf: This is a custom function named `fetch()`, which does not perform a request
-    await createResource(
-      {
-        url: screenshotID,
-        fetchDate: new Date(),
-        content: screenshot, // base64-encoded PNG image
-      },
-      60 * 60 * 24 // We only store the screenshot for 24 hours.
+const determineConcurrency = (req: Request): number => {
+  let concurrency: number = Number(req.query.concurrency ?? 1);
+  if (Number.isNaN(concurrency) || !Number.isSafeInteger(concurrency) || concurrency < 1) {
+    logger.warn(
+      PREFIX_SELENIUM +
+        chalk.yellowBright(`Invalid concurrency “${req.query.concurrency}” requested – using 1 fetcher only.`)
     );
-    return `For additional information, see https://${process.env.SUBDOMAIN ? process.env.SUBDOMAIN + "." : ""}${
-      process.env.DOMAIN
-    }/api/resource/${screenshotID}.`;
-  } catch (e) {
-    logger.warn(PREFIX_SELENIUM + chalk.yellowBright(`Unable to take screenshot “${screenshotID}”: ${e}`));
-    return "";
+    concurrency = 1;
   }
+  if (concurrency > Number(process.env.SELENIUM_MAX_CONCURRENCY)) {
+    logger.warn(
+      PREFIX_SELENIUM +
+        chalk.yellowBright(
+          `Desired concurrency “${concurrency}” is larger than the server allows – ` +
+            `using maximum value ${Number(process.env.SELENIUM_MAX_CONCURRENCY)} instead.`
+        )
+    );
+    concurrency = Number(process.env.SELENIUM_MAX_CONCURRENCY);
+  }
+  return concurrency;
 };
 
 /**
@@ -150,21 +88,21 @@ export class FetchController {
     accessRights: GENERAL_ACCESS + WRITE_STOCKS_ACCESS,
   })
   async fetchMorningstarData(req: Request, res: Response) {
-    let stocks: Stock[];
+    let stockList: Stock[];
 
     if (req.query.ticker) {
       // A single stock is requested.
       const ticker = req.query.ticker;
       if (typeof ticker === "string") {
-        stocks = [await readStock(ticker)];
-        if (!stocks[0].morningstarID) {
+        stockList = [await readStock(ticker)];
+        if (!stockList[0].morningstarID) {
           // If the only stock to use does not have a Morningstar ID, we throw an error.
           throw new APIError(404, `Stock ${ticker} does not have a Morningstar ID.`);
         }
       }
     } else {
       // When no specific stock is requested, we fetch all stocks from the database which have a Morningstar ID.
-      [stocks] = await readAllStocks({
+      [stockList] = await readAllStocks({
         where: {
           morningstarID: {
             not: null,
@@ -177,7 +115,7 @@ export class FetchController {
       });
     }
 
-    if (stocks.length === 0) {
+    if (stockList.length === 0) {
       // If no stocks are left, we return a 204 No Content response.
       res.status(204).end();
       return;
@@ -187,457 +125,18 @@ export class FetchController {
       res.status(202).end();
     }
 
-    const updatedStocks: Stock[] = [];
-    let successfulCount = 0;
-    let errorCount = 0;
-    let consecutiveErrorCount = 0;
-    const driver = await getDriver(true);
-    for await (const stock of stocks) {
-      if (
-        !req.query.noSkip &&
-        stock.morningstarLastFetch &&
-        // We only fetch stocks that have not been fetched in the last 12 hours.
-        new Date().getTime() - stock.morningstarLastFetch.getTime() < 1000 * 60 * 60 * 12
-      ) {
-        logger.info(
-          PREFIX_SELENIUM +
-            `Stock ${stock.ticker}: Skipping Morningstar fetch since last successful fetch was ${formatDistance(
-              stock.morningstarLastFetch.getTime(),
-              new Date().getTime(),
-              { addSuffix: true }
-            )}`
-        );
-        continue;
-      }
-      let industry: Industry = req.query.clear ? null : undefined;
-      let size: Size = req.query.clear ? null : undefined;
-      let style: Style = req.query.clear ? null : undefined;
-      let starRating: number = req.query.clear ? null : undefined;
-      let dividendYieldPercent: number = req.query.clear ? null : undefined;
-      let priceEarningRatio: number = req.query.clear ? null : undefined;
-      let currency: Currency = req.query.clear ? null : undefined;
-      let lastClose: number = req.query.clear ? null : undefined;
-      let morningstarFairValue: number = req.query.clear ? null : undefined;
-      let marketCap: number = req.query.clear ? null : undefined;
-      let low52w: number = req.query.clear ? null : undefined;
-      let high52w: number = req.query.clear ? null : undefined;
-      let description: string = req.query.clear ? null : undefined;
+    const stocks: FetcherWorkspace<Stock> = {
+      queued: [...stockList],
+      successful: [],
+      failed: [],
+    };
 
-      try {
-        await driver.get(
-          `https://tools.morningstar.it/it/stockreport/default.aspx?Site=us&id=${stock.morningstarID}` +
-            `&LanguageId=en-US&SecurityToken=${stock.morningstarID}]3]0]E0WWE$$ALL`
-        );
-        await driver.wait(
-          until.elementLocated(By.id("SnapshotContent")),
-          15000 // Wait for the page to load for a maximum of 15 seconds.
-        );
+    await Promise.all([...Array(determineConcurrency(req))].map(() => morningstarFetcher(req, stocks)));
 
-        // Prepare an error message header containing the stock name and ticker.
-        let errorMessage = `Error while fetching Morningstar data for ${stock.name} (${stock.ticker}):`;
-
-        try {
-          const industryString = (await driver.findElement(By.xpath(XPATH_INDUSTRY)).getText())
-            // Example: "Industry\nLumber & Wood Production"
-            .replace("Industry\n", "") // Remove headline
-            .replaceAll(/[^a-zA-Z0-9]/g, ""); // Remove all non-alphanumeric characters
-          if (isIndustry(industryString)) {
-            industry = industryString;
-          } else {
-            throw new TypeError(`Extracted industry “${industryString}” is no valid industry.`);
-          }
-        } catch (e) {
-          logger.warn(PREFIX_SELENIUM + chalk.yellowBright(`Stock ${stock.ticker}: Unable to extract industry: ${e}`));
-          if (stock.industry !== null) {
-            // If an industry for the stock is already stored in the database, but we cannot extract it now from the
-            // page, we log this as an error and send a message.
-            logger.error(
-              PREFIX_SELENIUM +
-                chalk.redBright(
-                  `Stock ${stock.ticker}: Extraction of industry failed unexpectedly. This incident will be reported.`
-                )
-            );
-            errorMessage += `\n\tUnable to extract industry: ${String(e.message).split(/[\n:{]/)[0]}`;
-          }
-        }
-
-        try {
-          const sizeAndStyle = (await driver.findElement(By.xpath(XPATH_SIZE_STYLE)).getText())
-            // Example: "Stock Style\nLarge-Blend"
-            .replace("Stock Style\n", "") // Remove headline
-            .split("-");
-          if (sizeAndStyle.length !== 2) {
-            throw new TypeError("No valid size and style available.");
-          }
-          if (isSize(sizeAndStyle[0])) {
-            size = sizeAndStyle[0];
-          } else {
-            throw new TypeError(`Extracted size “${sizeAndStyle[0]}” is no valid size.`);
-          }
-          if (isStyle(sizeAndStyle[1])) {
-            style = sizeAndStyle[1];
-          } else {
-            throw new TypeError(`Extracted style “${sizeAndStyle[1]}” is no valid style.`);
-          }
-        } catch (e) {
-          logger.warn(
-            PREFIX_SELENIUM + chalk.yellowBright(`Stock ${stock.ticker}: Unable to extract size and style: ${e}`)
-          );
-          if (stock.size !== null || stock.style !== null) {
-            // If size or style for the stock are already stored in the database, but we cannot extract them now from
-            // the page, we log this as an error and send a message.
-            logger.error(
-              PREFIX_SELENIUM +
-                chalk.redBright(
-                  `Stock ${stock.ticker}: Extraction of size and style failed unexpectedly. ` +
-                    `This incident will be reported.`
-                )
-            );
-            errorMessage += `\n\tUnable to extract size and style: ${String(e.message).split(/[\n:{]/)[0]}`;
-          }
-        }
-
-        try {
-          const starRatingString = (
-            await driver.findElement(By.xpath(XPATH_STAR_RATING)).getAttribute("alt")
-          ).replaceAll(/\D/g, ""); // Remove all non-digit characters
-          if (starRatingString.length === 0 || Number.isNaN(+starRatingString)) {
-            throw new TypeError(`Extracted star rating is no valid number.`);
-          }
-          starRating = +starRatingString;
-        } catch (e) {
-          logger.warn(
-            PREFIX_SELENIUM + chalk.yellowBright(`Stock ${stock.ticker}: Unable to extract star rating: ${e}`)
-          );
-          if (stock.starRating !== null) {
-            // If a star rating for the stock is already stored in the database, but we cannot extract it now from the
-            // page, we log this as an error and send a message.
-            logger.error(
-              PREFIX_SELENIUM +
-                chalk.redBright(
-                  `Stock ${stock.ticker}: Extraction of star rating failed unexpectedly. ` +
-                    `This incident will be reported.`
-                )
-            );
-            errorMessage += `\n\tUnable to extract star rating: ${String(e.message).split(/[\n:{]/)[0]}`;
-          }
-        }
-
-        try {
-          const dividendYieldPercentString = await driver.findElement(By.id("Col0Yield")).getText();
-          // Example: "2.1", or "-" if there is no dividend yield.
-          if (dividendYieldPercentString === "-") {
-            dividendYieldPercent = null;
-          } else {
-            if (dividendYieldPercentString.length === 0 || Number.isNaN(+dividendYieldPercentString)) {
-              throw new TypeError(`Extracted dividend yield is no valid number.`);
-            }
-            dividendYieldPercent = +dividendYieldPercentString;
-          }
-        } catch (e) {
-          logger.warn(
-            PREFIX_SELENIUM + chalk.yellowBright(`Stock ${stock.ticker}: Unable to extract dividend yield: ${e}`)
-          );
-          if (stock.dividendYieldPercent !== null) {
-            // If a dividend yield for the stock is already stored in the database, but we cannot extract it now from
-            // the page, we log this as an error and send a message.
-            logger.error(
-              PREFIX_SELENIUM +
-                chalk.redBright(
-                  `Stock ${stock.ticker}: Extraction of dividend yield failed unexpectedly. ` +
-                    `This incident will be reported.`
-                )
-            );
-            errorMessage += `\n\tUnable to extract dividend yield: ${String(e.message).split(/[\n:{]/)[0]}`;
-          }
-        }
-
-        try {
-          const priceEarningRatioString = (await driver.findElement(By.id("Col0PE")).getText()).replaceAll(",", "");
-          // Example: "20.5", "1,000" for larger numbers, or "-" if there is no P/E.
-          if (priceEarningRatioString === "-") {
-            priceEarningRatio = null;
-          } else {
-            if (priceEarningRatioString.length === 0 || Number.isNaN(+priceEarningRatioString)) {
-              throw new TypeError(`Extracted price earning ratio is no valid number.`);
-            }
-            priceEarningRatio = +priceEarningRatioString;
-          }
-        } catch (e) {
-          logger.warn(
-            PREFIX_SELENIUM + chalk.yellowBright(`Stock ${stock.ticker}: Unable to extract price earning ratio: ${e}`)
-          );
-          if (stock.priceEarningRatio !== null) {
-            // If a price earning ratio for the stock is already stored in the database, but we cannot extract it now
-            // from the page, we log this as an error and send a message.
-            logger.error(
-              PREFIX_SELENIUM +
-                chalk.redBright(
-                  `Stock ${stock.ticker}: Extraction of price earning ratio failed unexpectedly. ` +
-                    `This incident will be reported.`
-                )
-            );
-            errorMessage += `\n\tUnable to extract price earning ratio: ${String(e.message).split(/[\n:{]/)[0]}`;
-          }
-        }
-
-        try {
-          let currencyString = await driver.findElement(By.id("Col0PriceTime")).getText();
-          // Example: "17:35:38 CET | EUR  Minimum 15 Minutes Delay."
-          currencyString = currencyString.match(/\s+\|\s+([A-Z]{3})\s+/)[1];
-          if (isCurrency(currencyString)) {
-            currency = currencyString;
-          } else {
-            throw new TypeError(`Extracted currency code “${currencyString}” is no valid currency code.`);
-          }
-        } catch (e) {
-          logger.warn(PREFIX_SELENIUM + chalk.yellowBright(`Stock ${stock.ticker}: Unable to extract currency: ${e}`));
-          if (stock.currency !== null) {
-            // If a currency for the stock is already stored in the database, but we cannot extract it now from the
-            // page, we log this as an error and send a message.
-            logger.error(
-              PREFIX_SELENIUM +
-                chalk.redBright(
-                  `Stock ${stock.ticker}: Extraction of currency failed unexpectedly. This incident will be reported.`
-                )
-            );
-            errorMessage += `\n\tUnable to extract currency: ${String(e.message).split(/[\n:{]/)[0]}`;
-          }
-        }
-
-        try {
-          const lastCloseString = (await driver.findElement(By.id("Col0LastClose")).getText()).replaceAll(",", "");
-          // Example: "1,000.00", or "-" if there is no last close.
-          if (lastCloseString === "-") {
-            lastClose = null;
-          } else {
-            if (lastCloseString.length === 0 || Number.isNaN(+lastCloseString)) {
-              throw new TypeError(`Extracted last close is no valid number.`);
-            }
-            lastClose = +lastCloseString;
-          }
-        } catch (e) {
-          logger.warn(
-            PREFIX_SELENIUM + chalk.yellowBright(`Stock ${stock.ticker}: Unable to extract last close: ${e}`)
-          );
-          if (stock.lastClose !== null) {
-            // If a last close for the stock is already stored in the database, but we cannot extract it now from the
-            // page, we log this as an error and send a message.
-            logger.error(
-              PREFIX_SELENIUM +
-                chalk.redBright(
-                  `Stock ${stock.ticker}: Extraction of last close failed unexpectedly. This incident will be reported.`
-                )
-            );
-            errorMessage += `\n\tUnable to extract last close: ${String(e.message).split(/[\n:{]/)[0]}`;
-          }
-        }
-
-        try {
-          const morningstarFairValueString = (
-            await driver.findElement(By.xpath(XPATH_MORNINGSTAR_FAIR_VALUE)).getText()
-          ) // Example: "1,000.00 USD", or "-" if there is no Morningstar Fair Value.
-            .split(/\s+/)[0]
-            .replaceAll(",", "");
-          if (morningstarFairValueString === "-") {
-            morningstarFairValue = null;
-          } else {
-            if (morningstarFairValueString.length === 0 || Number.isNaN(+morningstarFairValueString)) {
-              throw new TypeError(`Extracted Morningstar Fair Value is no valid number.`);
-            }
-            morningstarFairValue = +morningstarFairValueString;
-          }
-        } catch (e) {
-          logger.warn(
-            PREFIX_SELENIUM +
-              chalk.yellowBright(`Stock ${stock.ticker}: Unable to extract Morningstar Fair Value: ${e}`)
-          );
-          if (stock.morningstarFairValue !== null) {
-            // If a Morningstar Fair Value for the stock is already stored in the database, but we cannot extract it
-            // now from the page, we log this as an error and send a message.
-            logger.error(
-              PREFIX_SELENIUM +
-                chalk.redBright(
-                  `Stock ${stock.ticker}: Extraction of Morningstar Fair Value failed unexpectedly. ` +
-                    `This incident will be reported.`
-                )
-            );
-            errorMessage += `\n\tUnable to extract Morningstar Fair Value: ${String(e.message).split(/[\n:{]/)[0]}`;
-          }
-        }
-
-        try {
-          const marketCapText = (await driver.findElement(By.id("Col0MCap")).getText())
-            // Example: "2,235.00Bil", or "-" if there is no market capitalization.
-            .replaceAll(",", "");
-          if (marketCapText === "-") {
-            marketCap = null;
-          } else {
-            if (marketCapText.includes("Bil")) {
-              marketCap = Math.round(1e9 * +marketCapText.substring(0, marketCapText.indexOf("Bil")));
-            } else if (marketCapText.includes("Mil")) {
-              marketCap = Math.round(1e6 * +marketCapText.substring(0, marketCapText.indexOf("Mil")));
-            } else {
-              marketCap = +marketCapText;
-            }
-            if (!marketCapText.match(/\d+/) || Number.isNaN(marketCap)) {
-              marketCap = req.query.clear ? null : undefined;
-              throw new TypeError(`Extracted market capitalization is no valid number.`);
-            }
-          }
-        } catch (e) {
-          logger.warn(
-            PREFIX_SELENIUM + chalk.yellowBright(`Stock ${stock.ticker}: Unable to extract Market Capitalization: ${e}`)
-          );
-          if (stock.marketCap !== null) {
-            // If a market capitalization for the stock is already stored in the database, but we cannot extract it now
-            // from the page, we log this as an error and send a message.
-            logger.error(
-              PREFIX_SELENIUM +
-                chalk.redBright(
-                  `Stock ${stock.ticker}: Extraction of Market Capitalization failed unexpectedly. ` +
-                    `This incident will be reported.`
-                )
-            );
-            errorMessage += `\n\tUnable to extract Market Capitalization: ${String(e.message).split(/[\n:{]/)[0]}`;
-          }
-        }
-
-        try {
-          const range52wStrings = (await driver.findElement(By.id("Col0WeekRange")).getText())
-            // Example: "1,000.00 - 2,000.00", or "-" if there is no 52 week price range.
-            .replaceAll(",", "")
-            .split(" - ");
-          if (range52wStrings[0] === "-") {
-            low52w = null;
-            high52w = null;
-          } else {
-            if (
-              range52wStrings.length !== 2 ||
-              range52wStrings[0].length === 0 ||
-              range52wStrings[1].length === 0 ||
-              Number.isNaN(+range52wStrings[0]) ||
-              Number.isNaN(+range52wStrings[1])
-            ) {
-              throw new TypeError(`Extracted 52 week low or high is no valid number.`);
-            }
-            low52w = +range52wStrings[0];
-            high52w = +range52wStrings[1];
-          }
-        } catch (e) {
-          logger.warn(
-            PREFIX_SELENIUM + chalk.yellowBright(`Stock ${stock.ticker}: Unable to extract 52 week price range: ${e}`)
-          );
-          if (stock.low52w !== null || stock.high52w !== null) {
-            // If a 52 week price range for the stock is already stored in the database, but we cannot extract it now
-            // from the page, we log this as an error and send a message.
-            logger.error(
-              PREFIX_SELENIUM +
-                chalk.redBright(
-                  `Stock ${stock.ticker}: Extraction of 52 week price range failed unexpectedly. ` +
-                    `This incident will be reported.`
-                )
-            );
-            errorMessage += `\n\tUnable to extract 52 week price range: ${String(e.message).split(/[\n:{]/)[0]}`;
-          }
-        }
-
-        try {
-          description = await driver.findElement(By.xpath(XPATH_DESCRIPTION)).getText();
-        } catch (e) {
-          logger.warn(
-            PREFIX_SELENIUM + chalk.yellowBright(`Stock ${stock.ticker}: Unable to extract description: ${e}`)
-          );
-          if (stock.description !== null) {
-            // If a description for the stock is already stored in the database, but we cannot extract it now from
-            // the page, we log this as an error and send a message.
-            logger.error(
-              PREFIX_SELENIUM +
-                chalk.redBright(
-                  `Stock ${stock.ticker}: Extraction of description failed unexpectedly. ` +
-                    `This incident will be reported.`
-                )
-            );
-            errorMessage += `\n\tUnable to extract description: ${String(e.message).split(/[\n:{]/)[0]}`;
-          }
-        }
-
-        if (errorMessage.includes("\n")) {
-          // An error occurred if and only if the error message contains a newline character.
-          // We take a screenshot and send a message.
-          errorMessage += `\n${await takeScreenshot(driver, stock, "morningstar")}`;
-          await signal.sendMessage(SIGNAL_PREFIX_ERROR + errorMessage, "fetchError");
-          await new Promise((resolve) => setTimeout(resolve, 3000)); // Cool down for 3 seconds.
-          errorCount += 1;
-          consecutiveErrorCount += 1;
-        } else {
-          successfulCount += 1;
-          consecutiveErrorCount = 0;
-        }
-        // Update the stock in the database.
-        await updateStock(stock.ticker, {
-          industry,
-          size,
-          style,
-          morningstarLastFetch: errorMessage.includes("\n") ? undefined : new Date(),
-          starRating,
-          dividendYieldPercent,
-          priceEarningRatio,
-          currency,
-          lastClose,
-          morningstarFairValue,
-          marketCap,
-          low52w,
-          high52w,
-          description,
-        });
-        updatedStocks.push(await readStock(stock.ticker));
-      } catch (e) {
-        errorCount += 1;
-        consecutiveErrorCount += 1;
-        if (req.query.ticker) {
-          // If the request was for a single stock, we shut down the driver and throw an error.
-          await quitDriver(driver);
-          throw new APIError(
-            502,
-            `Stock ${stock.ticker}: Unable to fetch Morningstar data: ${String(e.message).split(/[\n:{]/)[0]}`
-          );
-        }
-        logger.error(
-          PREFIX_SELENIUM + chalk.redBright(`Stock ${stock.ticker}: Unable to fetch Morningstar data: ${e}`)
-        );
-        await signal.sendMessage(
-          SIGNAL_PREFIX_ERROR +
-            `Stock ${stock.ticker}: Unable to fetch Morningstar data: ${
-              String(e.message).split(/[\n:{]/)[0]
-            }\n${await takeScreenshot(driver, stock, "morningstar")}`,
-          "fetchError"
-        );
-        await new Promise((resolve) => setTimeout(resolve, 3000)); // Cool down for 3 seconds.
-      }
-      if (consecutiveErrorCount >= 10) {
-        // If we have 10 consecutive errors, we stop fetching data, since something is probably wrong.
-        logger.error(
-          PREFIX_SELENIUM +
-            chalk.redBright(
-              `Aborting fetching information from Morningstar after ${consecutiveErrorCount} consecutive failures, ` +
-                `${successfulCount} successful fetches and ${errorCount} total failures. Will continue next time.`
-            )
-        );
-        await signal.sendMessage(
-          SIGNAL_PREFIX_ERROR +
-            `Aborting fetching information from Morningstar after ${consecutiveErrorCount} consecutive failures, ` +
-            `${successfulCount} successful fetches and ${errorCount} total failures. Will continue next time.`,
-          "fetchError"
-        );
-        break;
-      }
-    }
-    await quitDriver(driver);
-    if (updatedStocks.length === 0) {
+    if (stocks.successful.length === 0) {
       res.status(204).end();
     } else {
-      res.status(200).json(updatedStocks).end();
+      res.status(200).json(stocks.successful).end();
     }
   }
 
@@ -654,21 +153,21 @@ export class FetchController {
     accessRights: GENERAL_ACCESS + WRITE_STOCKS_ACCESS,
   })
   async fetchMarketScreenerData(req: Request, res: Response) {
-    let stocks: Stock[];
+    let stockList: Stock[];
 
     if (req.query.ticker) {
       // A single stock is requested.
       const ticker = req.query.ticker;
       if (typeof ticker === "string") {
-        stocks = [await readStock(ticker)];
-        if (!stocks[0].marketScreenerID) {
+        stockList = [await readStock(ticker)];
+        if (!stockList[0].marketScreenerID) {
           // If the only stock to use does not have a MarketScreener ID, we throw an error.
           throw new APIError(404, `Stock ${ticker} does not have a MarketScreener ID.`);
         }
       }
     } else {
       // When no specific stock is requested, we fetch all stocks from the database which have a Market Screener ID.
-      [stocks] = await readAllStocks({
+      [stockList] = await readAllStocks({
         where: {
           marketScreenerID: {
             not: null,
@@ -681,7 +180,7 @@ export class FetchController {
       });
     }
 
-    if (stocks.length === 0) {
+    if (stockList.length === 0) {
       // If no stocks are left, we return a 204 No Content response.
       res.status(204).end();
       return;
@@ -691,216 +190,18 @@ export class FetchController {
       res.status(202).end();
     }
 
-    const updatedStocks: Stock[] = [];
-    let successfulCount = 0;
-    let errorCount = 0;
-    let consecutiveErrorCount = 0;
-    const driver = await getDriver(true);
-    for await (const stock of stocks) {
-      if (
-        !req.query.noSkip &&
-        stock.marketScreenerLastFetch &&
-        // We only fetch stocks that have not been fetched in the last 12 hours.
-        new Date().getTime() - stock.marketScreenerLastFetch.getTime() < 1000 * 60 * 60 * 12
-      ) {
-        logger.info(
-          PREFIX_SELENIUM +
-            `Stock ${stock.ticker}: Skipping MarketScreener fetch because last fetch was ${formatDistance(
-              stock.marketScreenerLastFetch.getTime(),
-              new Date().getTime(),
-              { addSuffix: true }
-            )}`
-        );
-        continue;
-      }
-      let analystConsensus: number = req.query.clear ? null : undefined;
-      let analystCount: number = req.query.clear ? null : undefined;
-      let analystTargetPrice: number = req.query.clear ? null : undefined;
+    const stocks: FetcherWorkspace<Stock> = {
+      queued: [...stockList],
+      successful: [],
+      failed: [],
+    };
 
-      try {
-        await driver.get(`https://www.marketscreener.com/quote/stock/${stock.marketScreenerID}/`);
-        // Wait for the page to load for a maximum of 5 seconds.
-        await driver.wait(until.elementLocated(By.id("zbCenter")), 5000);
+    await Promise.all([...Array(determineConcurrency(req))].map(() => marketScreenerFetcher(req, stocks)));
 
-        // Prepare an error message header containing the stock name and ticker.
-        let errorMessage = `Error while fetching MarketScreener data for stock ${stock.ticker}:`;
-
-        try {
-          // Locate the div containing all relevant analyst-related information.
-          const consensusTableDiv = await driver.findElement(By.xpath(XPATH_CONSENSUS_DIV));
-
-          try {
-            const analystConsensusMatches = (
-              await (await consensusTableDiv.findElement(By.xpath(XPATH_CONSENSUS_NOTE))).getAttribute("title")
-            ) // Example: " Note : 9.1 / 10"
-              .match(/(\d+(\.\d+)?)/g); // Extract the first decimal number from the title.
-            if (
-              analystConsensusMatches === null ||
-              analystConsensusMatches.length < 1 ||
-              Number.isNaN(+analystConsensusMatches[0])
-            ) {
-              throw new TypeError(`Extracted analyst consensus is no valid number.`);
-            }
-            analystConsensus = +analystConsensusMatches[0];
-          } catch (e) {
-            logger.warn(
-              PREFIX_SELENIUM + chalk.yellowBright(`Stock ${stock.ticker}: Unable to extract Analyst Consensus: ${e}`)
-            );
-            if (stock.analystConsensus !== null) {
-              // If an analyst consensus is already stored in the database, but we cannot extract it from the page, we
-              // log this as an error and send a message.
-              logger.error(
-                PREFIX_SELENIUM +
-                  chalk.redBright(
-                    `Stock ${stock.ticker}: Extraction of analyst consensus failed unexpectedly. ` +
-                      `This incident will be reported.`
-                  )
-              );
-              errorMessage += `\n\tUnable to extract Analyst Consensus: ${String(e.message).split(/[\n:{]/)[0]}`;
-            }
-          }
-
-          try {
-            analystCount = +(await (await consensusTableDiv.findElement(By.xpath(XPATH_ANALYST_COUNT))).getText());
-          } catch (e) {
-            logger.warn(
-              PREFIX_SELENIUM + chalk.yellowBright(`Stock ${stock.ticker}: Unable to extract Analyst Count: ${e}`)
-            );
-            if (stock.analystCount !== null) {
-              // If an analyst count is already stored in the database, but we cannot extract it from the page, we log
-              // this as an error and send a message.
-              logger.error(
-                PREFIX_SELENIUM +
-                  chalk.redBright(
-                    `Stock ${stock.ticker}: Extraction of analyst count failed unexpectedly. ` +
-                      `This incident will be reported.`
-                  )
-              );
-              errorMessage += `\n\tUnable to extract Analyst Count: ${String(e.message).split(/[\n:{]/)[0]}`;
-            }
-          }
-
-          try {
-            // We need the last close price to calculate the analyst target price.
-            if (!stock.lastClose) {
-              throw new Error("No Last Close price available to compare spread against.");
-            }
-            const analystTargetPriceMatches = (
-              await (await consensusTableDiv.findElement(By.xpath(XPATH_SPREAD_AVERAGE_TARGET))).getText()
-            )
-              .replaceAll(",", ".")
-              .match(/(\-)?\d+(\.\d+)?/g);
-            if (
-              analystTargetPriceMatches === null ||
-              analystTargetPriceMatches.length !== 1 ||
-              Number.isNaN(+analystTargetPriceMatches[0])
-            ) {
-              throw new TypeError(`Extracted analyst target price is no valid number.`);
-            }
-            analystTargetPrice = stock.lastClose * (+analystTargetPriceMatches[0] / 100 + 1);
-          } catch (e) {
-            logger.warn(
-              PREFIX_SELENIUM +
-                chalk.yellowBright(`Stock ${stock.ticker}: Unable to extract Analyst Target Price: ${e}`)
-            );
-            if (stock.analystTargetPrice !== null) {
-              // If an analyst target price is already stored in the database, but we cannot extract it from the page,
-              // we log this as an error and send a message.
-              logger.error(
-                PREFIX_SELENIUM +
-                  chalk.redBright(
-                    `Stock ${stock.ticker}: Extraction of analyst target price failed unexpectedly. ` +
-                      `This incident will be reported.`
-                  )
-              );
-              errorMessage += `\n\tUnable to extract Analyst Target Price: ${String(e.message).split(/[\n:{]/)[0]}`;
-            }
-          }
-        } catch (e) {
-          logger.warn(
-            PREFIX_SELENIUM +
-              chalk.yellowBright(`Stock ${stock.ticker}: \n\tUnable to extract Analyst Information: ${e}`)
-          );
-          if (stock.analystConsensus !== null || stock.analystCount !== null || stock.analystTargetPrice !== null) {
-            // If any of the analyst-related information is already stored in the database, but we cannot extract it
-            // from the page, we log this as an error and send a message.
-            logger.error(
-              PREFIX_SELENIUM +
-                chalk.redBright(
-                  `Stock ${stock.ticker}: Extraction of analyst information failed unexpectedly. ` +
-                    `This incident will be reported.`
-                )
-            );
-            errorMessage += `\n\tUnable to extract Analyst Information: ${String(e.message).split(/[\n:{]/)[0]}`;
-          }
-        }
-
-        if (errorMessage.includes("\n")) {
-          // An error occurred if and only if the error message contains a newline character.
-          // We take a screenshot and send a message.
-          errorMessage += `\n${await takeScreenshot(driver, stock, "marketscreener")}`;
-          await signal.sendMessage(SIGNAL_PREFIX_ERROR + errorMessage, "fetchError");
-          await new Promise((resolve) => setTimeout(resolve, 3000)); // Cool down for 3 seconds.
-          errorCount += 1;
-          consecutiveErrorCount += 1;
-        } else {
-          successfulCount += 1;
-          consecutiveErrorCount = 0;
-        }
-        // Update the stock in the database.
-        await updateStock(stock.ticker, {
-          marketScreenerLastFetch: errorMessage.includes("\n") ? undefined : new Date(),
-          analystConsensus,
-          analystCount,
-          analystTargetPrice,
-        });
-        updatedStocks.push(await readStock(stock.ticker));
-      } catch (e) {
-        errorCount += 1;
-        consecutiveErrorCount += 1;
-        if (req.query.ticker) {
-          // If this request was for a single stock, we shut down the driver and throw an error.
-          await quitDriver(driver);
-          throw new APIError(
-            502,
-            `Stock ${stock.ticker}: Unable to fetch MarketScreener data: ${String(e.message).split(/[\n:{]/)[0]}`
-          );
-        }
-        logger.error(
-          PREFIX_SELENIUM + chalk.redBright(`Stock ${stock.ticker}: Unable to fetch MarketScreener data: ${e}`)
-        );
-        await signal.sendMessage(
-          SIGNAL_PREFIX_ERROR +
-            `Stock ${stock.ticker}: Unable to fetch MarketScreener data: ${
-              String(e.message).split(/[\n:{]/)[0]
-            }\n${await takeScreenshot(driver, stock, "marketscreener")}`,
-          "fetchError"
-        );
-        await new Promise((resolve) => setTimeout(resolve, 3000)); // Cool down for 3 seconds.
-      }
-      if (consecutiveErrorCount >= 5) {
-        // If we have 5 consecutive errors, we stop fetching data, since something is probably wrong.
-        logger.error(
-          PREFIX_SELENIUM +
-            chalk.redBright(
-              `Aborting fetching information from MarketScreener after ${consecutiveErrorCount} consecutive failures,` +
-                ` ${successfulCount} successful fetches and ${errorCount} total failures. Will continue next time.`
-            )
-        );
-        await signal.sendMessage(
-          SIGNAL_PREFIX_ERROR +
-            `Aborting fetching information from MarketScreener after ${consecutiveErrorCount} consecutive failures, ` +
-            `${successfulCount} successful fetches and ${errorCount} total failures. Will continue next time.`,
-          "fetchError"
-        );
-        break;
-      }
-    }
-    await quitDriver(driver);
-    if (updatedStocks.length === 0) {
+    if (stocks.successful.length === 0) {
       res.status(204).end();
     } else {
-      res.status(200).json(updatedStocks).end();
+      res.status(200).json(stocks.successful).end();
     }
   }
 
@@ -917,21 +218,21 @@ export class FetchController {
     accessRights: GENERAL_ACCESS + WRITE_STOCKS_ACCESS,
   })
   async fetchMSCIData(req: Request, res: Response) {
-    let stocks: Stock[];
+    let stockList: Stock[];
 
     if (req.query.ticker) {
       // A single stock is requested.
       const ticker = req.query.ticker;
       if (typeof ticker === "string") {
-        stocks = [await readStock(ticker)];
-        if (!stocks[0].msciID) {
+        stockList = [await readStock(ticker)];
+        if (!stockList[0].msciID) {
           // If the only stock to use does not have an MSCI ID, we throw an error.
           throw new APIError(404, `Stock ${ticker} does not have an MSCI ID.`);
         }
       }
     } else {
       // When no specific stock is requested, we fetch all stocks from the database which have an MSCI ID.
-      [stocks] = await readAllStocks({
+      [stockList] = await readAllStocks({
         where: {
           msciID: {
             not: null,
@@ -944,7 +245,7 @@ export class FetchController {
       });
     }
 
-    if (stocks.length === 0) {
+    if (stockList.length === 0) {
       // If no stocks are left, we return a 204 No Content response.
       res.status(204).end();
       return;
@@ -954,192 +255,18 @@ export class FetchController {
       res.status(202).end();
     }
 
-    const updatedStocks: Stock[] = [];
-    let successfulCount = 0;
-    let errorCount = 0;
-    let consecutiveErrorCount = 0;
-    const driver = await getDriver();
-    for await (const stock of stocks) {
-      if (
-        !req.query.noSkip &&
-        stock.msciLastFetch &&
-        // We only fetch stocks that have not been fetched in the last 14 days.
-        new Date().getTime() - stock.msciLastFetch.getTime() < 1000 * 60 * 60 * 24 * 14
-      ) {
-        logger.info(
-          PREFIX_SELENIUM +
-            `Stock ${stock.ticker}: Skipping MSCI fetch since last successful fetch was ${formatDistance(
-              stock.msciLastFetch.getTime(),
-              new Date().getTime(),
-              { addSuffix: true }
-            )}`
-        );
-        continue;
-      }
-      if (successfulCount >= 50) {
-        // If we have fetched 50 stocks successfully, we stop fetching data to avoid rate limiting.
-        logger.info(
-          PREFIX_SELENIUM +
-            chalk.greenBright(
-              `Successfully fetched MSCI information for ${successfulCount} stocks (${errorCount} errors). ` +
-                `Pausing now to avoid rate limiting. Will continue next time.`
-            )
-        );
-        await signal.sendMessage(
-          SIGNAL_PREFIX_INFO +
-            `Successfully fetched MSCI information for ${successfulCount} stocks (${errorCount} errors). ` +
-            `Pausing now to avoid rate limiting. Will continue next time.`,
-          "fetchError"
-        );
-        break;
-      }
-      let msciESGRating: MSCIESGRating = req.query.clear ? null : undefined;
-      let msciTemperature: number = req.query.clear ? null : undefined;
+    const stocks: FetcherWorkspace<Stock> = {
+      queued: [...stockList],
+      successful: [],
+      failed: [],
+    };
 
-      try {
-        await driver.manage().deleteAllCookies(); // Delete all cookies since MSCI allows only 4 requests per session.
-        if (successfulCount > 0 || errorCount > 0) {
-          // Wait 15 seconds between requests to avoid rate limiting.
-          await new Promise((resolve) => setTimeout(resolve, 15000));
-        }
-        await driver.get(
-          `https://www.msci.com/our-solutions/esg-investing/esg-ratings-climate-search-tool/issuer/${stock.msciID}`
-        );
-        await driver.wait(
-          until.elementsLocated(By.className("esg-expandable")),
-          10000 // Wait for the page to load for a maximum of 10 seconds.
-        );
+    await Promise.all([...Array(determineConcurrency(req))].map(() => msciFetcher(req, stocks)));
 
-        // Prepare an error message header containing the stock name and ticker.
-        let errorMessage = `Error while fetching MSCI information for ${stock.name} (${stock.ticker}):`;
-
-        try {
-          const esgClassName = await driver
-            .findElement(By.className("ratingdata-company-rating"))
-            .getAttribute("class"); // Example: "esg-rating-circle-bbb"
-          const msciESGRatingString = esgClassName.substring(esgClassName.lastIndexOf("-") + 1).toUpperCase();
-          if (isMSCIESGRating(msciESGRatingString)) {
-            msciESGRating = msciESGRatingString;
-          } else {
-            throw new TypeError(`Extracted MSCI ESG Rating “${msciESGRatingString}” is no valid MSCI ESG Rating.`);
-          }
-        } catch (e) {
-          logger.warn(
-            PREFIX_SELENIUM + chalk.yellowBright(`Stock ${stock.ticker}: Unable to extract MSCI ESG Rating: ${e}`)
-          );
-          if (stock.msciESGRating !== null) {
-            // If an MSCI ESG Rating is already stored in the database, but we cannot extract it from the page, we log
-            // this as an error and send a message.
-            logger.error(
-              PREFIX_SELENIUM +
-                chalk.redBright(
-                  `Stock ${stock.ticker}: Extraction of MSCI ESG Rating failed unexpectedly. ` +
-                    `This incident will be reported.`
-                )
-            );
-            errorMessage += `\n\tUnable to extract MSCI ESG Rating: ${String(e.message).split(/[\n:{]/)[0]}`;
-          }
-        }
-
-        try {
-          const msciTemperatureMatches = (
-            await driver.findElement(By.className("implied-temp-rise-value")).getAttribute("outerText")
-          ) // Example: "2.5°C"
-            .match(/(\d+(\.\d+)?)/g);
-          if (
-            msciTemperatureMatches === null ||
-            msciTemperatureMatches.length !== 1 ||
-            Number.isNaN(+msciTemperatureMatches[0])
-          ) {
-            throw new TypeError(`Extracted MSCI Implied Temperature Rise is no valid number.`);
-          }
-          msciTemperature = +msciTemperatureMatches[0];
-        } catch (e) {
-          logger.warn(
-            PREFIX_SELENIUM +
-              chalk.yellowBright(`Stock ${stock.ticker}: Unable to extract MSCI Implied Temperature Rise: ${e}`)
-          );
-          if (stock.msciTemperature !== null) {
-            // If an MSCI Implied Temperature Rise is already stored in the database, but we cannot extract it from the
-            // page, we log this as an error and send a message.
-            logger.error(
-              PREFIX_SELENIUM +
-                chalk.redBright(
-                  `Stock ${stock.ticker}: Extraction of MSCI Implied Temperature Rise failed unexpectedly. ` +
-                    `This incident will be reported.`
-                )
-            );
-            errorMessage += `\n\tUnable to extract MSCI Implied Temperature Rise: ${
-              String(e.message).split(/[\n:{]/)[0]
-            }`;
-          }
-        }
-
-        if (errorMessage.includes("\n")) {
-          // An error occurred if and only if the error message contains a newline character.
-          // We take a screenshot and send a message.
-          errorMessage += `\n${await takeScreenshot(driver, stock, "msci")}`;
-          await signal.sendMessage(SIGNAL_PREFIX_ERROR + errorMessage, "fetchError");
-          await new Promise((resolve) => setTimeout(resolve, 3000)); // Cool down for 3 seconds.
-          errorCount += 1;
-          consecutiveErrorCount += 1;
-        } else {
-          successfulCount += 1;
-          consecutiveErrorCount = 0;
-        }
-        // Update the stock in the database.
-        await updateStock(stock.ticker, {
-          msciLastFetch: errorMessage.includes("\n") ? undefined : new Date(),
-          msciESGRating,
-          msciTemperature,
-        });
-        updatedStocks.push(await readStock(stock.ticker));
-      } catch (e) {
-        errorCount += 1;
-        consecutiveErrorCount += 1;
-        if (req.query.ticker) {
-          // If this request was for a single stock, we shut down the driver and throw an error.
-          await quitDriver(driver);
-          throw new APIError(
-            502,
-            `Stock ${stock.ticker}: Unable to fetch MSCI information: ${String(e.message).split(/[\n:{]/)[0]}`
-          );
-        }
-        logger.error(
-          PREFIX_SELENIUM + chalk.redBright(`Stock ${stock.ticker}: Unable to fetch MSCI information: ${e}`)
-        );
-        await signal.sendMessage(
-          SIGNAL_PREFIX_ERROR +
-            `Stock ${stock.ticker}: Unable to fetch MSCI information: ${
-              String(e.message).split(/[\n:{]/)[0]
-            }\n${await takeScreenshot(driver, stock, "msci")}`,
-          "fetchError"
-        );
-        await new Promise((resolve) => setTimeout(resolve, 3000)); // Cool down for 3 seconds.
-      }
-      if (consecutiveErrorCount >= 10) {
-        // If we have 10 consecutive errors, we stop fetching data, since something is probably wrong.
-        logger.error(
-          PREFIX_SELENIUM +
-            chalk.redBright(
-              `Aborting fetching information from MSCI after ${consecutiveErrorCount} consecutive failures, ` +
-                `${successfulCount} successful fetches and ${errorCount} total failures. Will continue next time.`
-            )
-        );
-        await signal.sendMessage(
-          SIGNAL_PREFIX_ERROR +
-            `Aborting fetching information from MSCI after ${consecutiveErrorCount} consecutive failures, ` +
-            `${successfulCount} successful fetches and ${errorCount} total failures. Will continue next time.`,
-          "fetchError"
-        );
-        break;
-      }
-    }
-    await quitDriver(driver);
-    if (updatedStocks.length === 0) {
+    if (stocks.successful.length === 0) {
       res.status(204).end();
     } else {
-      res.status(200).json(updatedStocks);
+      res.status(200).json(stocks.successful);
     }
   }
 
@@ -1156,21 +283,21 @@ export class FetchController {
     accessRights: GENERAL_ACCESS + WRITE_STOCKS_ACCESS,
   })
   async fetchRefinitivData(req: Request, res: Response) {
-    let stocks: Stock[];
+    let stockList: Stock[];
 
     if (req.query.ticker) {
       // A single stock is requested.
       const ticker = req.query.ticker;
       if (typeof ticker === "string") {
-        stocks = [await readStock(ticker)];
-        if (!stocks[0].ric) {
+        stockList = [await readStock(ticker)];
+        if (!stockList[0].ric) {
           // If the only stock to use does not have a RIC, we throw an error.
           throw new APIError(404, `Stock ${ticker} does not have a RIC.`);
         }
       }
     } else {
       // When no specific stock is requested, we fetch all stocks from the database which have a RIC.
-      [stocks] = await readAllStocks({
+      [stockList] = await readAllStocks({
         where: {
           ric: {
             not: null,
@@ -1183,7 +310,7 @@ export class FetchController {
       });
     }
 
-    if (stocks.length === 0) {
+    if (stockList.length === 0) {
       // If no stocks are left, we return a 204 No Content response.
       res.status(204).end();
       return;
@@ -1193,177 +320,18 @@ export class FetchController {
       res.status(202).end();
     }
 
-    const updatedStocks: Stock[] = [];
-    let successfulCount = 0;
-    let errorCount = 0;
-    let consecutiveErrorCount = 0;
-    const driver = await getDriver(true);
-    for await (const stock of stocks) {
-      if (
-        !req.query.noSkip &&
-        stock.refinitivLastFetch &&
-        // We only fetch stocks that have not been fetched in the last 7 days.
-        new Date().getTime() - stock.refinitivLastFetch.getTime() < 1000 * 60 * 60 * 24 * 7
-      ) {
-        logger.info(
-          PREFIX_SELENIUM +
-            `Stock ${stock.ticker}: Skipping Refinitiv fetch because last fetch was ${formatDistance(
-              stock.refinitivLastFetch.getTime(),
-              new Date().getTime(),
-              { addSuffix: true }
-            )}`
-        );
-        continue;
-      }
-      let refinitivESGScore: number = req.query.clear ? null : undefined;
-      let refinitivEmissions: number = req.query.clear ? null : undefined;
+    const stocks: FetcherWorkspace<Stock> = {
+      queued: [...stockList],
+      successful: [],
+      failed: [],
+    };
 
-      try {
-        // Delete all cookies since Refinitiv allows only 100 requests per session.
-        await driver.manage().deleteAllCookies();
-        await driver.get(`https://www.refinitiv.com/bin/esg/esgsearchresult?ricCode=${stock.ric}`);
-        // Wait for the page to load for a maximum of 5 seconds.
-        await driver.wait(until.elementLocated(By.css("pre")), 5000);
+    await Promise.all([...Array(determineConcurrency(req))].map(() => refinitivFetcher(req, stocks)));
 
-        // Extract the JSON content from the page.
-        const refinitivJSONText = await (await driver.findElement(By.css("pre"))).getText();
-
-        if (refinitivJSONText === "{}") {
-          throw new APIError(502, "No Refinitiv information available.");
-        }
-
-        const refinitivJSON = JSON.parse(refinitivJSONText);
-
-        if (refinitivJSON.status && refinitivJSON.status.limitExceeded === true) {
-          // If the limit has been exceeded, we stop fetching data and throw an error.
-          throw new APIError(429, "Limit exceeded.");
-        }
-
-        // Prepare an error message header containing the stock name and ticker.
-        let errorMessage = `Error while fetching Refinitiv information for stock ${stock.ticker}:`;
-
-        try {
-          refinitivESGScore = +refinitivJSON.esgScore["TR.TRESG"].score;
-        } catch (e) {
-          logger.warn(
-            PREFIX_SELENIUM + chalk.yellowBright(`Stock ${stock.ticker}: Unable to extract Refinitiv ESG Score: ${e}`)
-          );
-          if (stock.refinitivESGScore !== null) {
-            // If a Refinitiv ESG Score is already stored in the database, but we cannot extract it from the page, we
-            // log this as an error and send a message.
-            logger.error(
-              PREFIX_SELENIUM +
-                chalk.redBright(
-                  `Stock ${stock.ticker}: Extraction of Refinitiv ESG Score failed unexpectedly. ` +
-                    `This incident will be reported.`
-                )
-            );
-            errorMessage += `\n\tUnable to extract Refinitiv ESG Score: ${String(e.message).split(/[\n:{]/)[0]}`;
-          }
-        }
-
-        try {
-          refinitivEmissions = +refinitivJSON.esgScore["TR.TRESGEmissions"].score;
-        } catch (e) {
-          logger.warn(
-            PREFIX_SELENIUM + chalk.yellowBright(`Stock ${stock.ticker}: Unable to extract Refinitiv Emissions: ${e}`)
-          );
-          if (stock.refinitivEmissions !== null) {
-            // If a Refinitiv Emissions Rating is already stored in the database, but we cannot extract it from the
-            // page, we log this as an error and send a message.
-            logger.error(
-              PREFIX_SELENIUM +
-                chalk.redBright(
-                  `Stock ${stock.ticker}: Extraction of Refinitiv Emissions failed unexpectedly. ` +
-                    `This incident will be reported.`
-                )
-            );
-            errorMessage += `\n\tUnable to extract Refinitiv Emissions: ${String(e.message).split(/[\n:{]/)[0]}`;
-          }
-        }
-
-        if (errorMessage.includes("\n")) {
-          // An error occurred if and only if the error message contains a newline character.
-          // We take a screenshot and send a message.
-          errorMessage += `\n${await takeScreenshot(driver, stock, "refinitiv")}`;
-          await signal.sendMessage(SIGNAL_PREFIX_ERROR + errorMessage, "fetchError");
-          await new Promise((resolve) => setTimeout(resolve, 3000)); // Cool down for 3 seconds.
-          errorCount += 1;
-          consecutiveErrorCount += 1;
-        } else {
-          successfulCount += 1;
-          consecutiveErrorCount = 0;
-        }
-        // Update the stock in the database.
-        await updateStock(stock.ticker, {
-          refinitivLastFetch: errorMessage.includes("\n") ? undefined : new Date(),
-          refinitivESGScore,
-          refinitivEmissions,
-        });
-        updatedStocks.push(await readStock(stock.ticker));
-      } catch (e) {
-        errorCount += 1;
-        consecutiveErrorCount += 1;
-        if (req.query.ticker) {
-          // If this request was for a single stock, we shut down the driver and throw an error.
-          await quitDriver(driver);
-          throw new APIError(
-            (e as Error).message.includes("Limit exceeded") ? 429 : 502,
-            `Stock ${stock.ticker}: Unable to fetch Refinitiv information: ${String(e.message).split(/[\n:{]/)[0]}`
-          );
-        }
-        if ((e as Error).message.includes("Limit exceeded")) {
-          // If the limit has been exceeded, we stop fetching data and log an error.
-          logger.error(
-            PREFIX_SELENIUM +
-              chalk.redBright(
-                `Aborting fetching information from Refinitiv after exceeding limit ` +
-                  `(${successfulCount} successful fetches). Will continue next time.`
-              )
-          );
-          await signal.sendMessage(
-            SIGNAL_PREFIX_ERROR +
-              `Aborting fetching information from Refinitiv after exceeding limit ` +
-              `(${successfulCount} successful fetches). Will continue next time.`,
-            "fetchError"
-          );
-          break;
-        }
-        logger.error(
-          PREFIX_SELENIUM + chalk.redBright(`Stock ${stock.ticker}: Unable to fetch Refinitiv information: ${e}`)
-        );
-        await signal.sendMessage(
-          SIGNAL_PREFIX_ERROR +
-            `Stock ${stock.ticker}: Unable to fetch Refinitiv information: ${
-              String(e.message).split(/[\n:{]/)[0]
-            }\n${await takeScreenshot(driver, stock, "refinitiv")}`,
-          "fetchError"
-        );
-        await new Promise((resolve) => setTimeout(resolve, 3000)); // Cool down for 3 seconds.
-      }
-      if (consecutiveErrorCount >= 5) {
-        // If we have 5 consecutive errors, we stop fetching data, since something is probably wrong.
-        logger.error(
-          PREFIX_SELENIUM +
-            chalk.redBright(
-              `Aborting fetching information from Refinitiv after ${consecutiveErrorCount} consecutive failures, ` +
-                `${successfulCount} successful fetches and ${errorCount} total failures. Will continue next time.`
-            )
-        );
-        await signal.sendMessage(
-          SIGNAL_PREFIX_ERROR +
-            `Aborting fetching information from Refinitiv after ${consecutiveErrorCount} consecutive failures, ` +
-            `${successfulCount} successful fetches and ${errorCount} total failures. Will continue next time.`,
-          "fetchError"
-        );
-        break;
-      }
-    }
-    await quitDriver(driver);
-    if (updatedStocks.length === 0) {
+    if (stocks.successful.length === 0) {
       res.status(204).end();
     } else {
-      res.status(200).json(updatedStocks).end();
+      res.status(200).json(stocks.successful).end();
     }
   }
 
@@ -1380,21 +348,21 @@ export class FetchController {
     accessRights: GENERAL_ACCESS + WRITE_STOCKS_ACCESS,
   })
   async fetchSPData(req: Request, res: Response) {
-    let stocks: Stock[];
+    let stockList: Stock[];
 
     if (req.query.ticker) {
       // A single stock is requested.
       const ticker = req.query.ticker;
       if (typeof ticker === "string") {
-        stocks = [await readStock(ticker)];
-        if (!stocks[0].spID) {
+        stockList = [await readStock(ticker)];
+        if (!stockList[0].spID) {
           // If the only stock to use does not have an S&P ID, we throw an error.
           throw new APIError(404, `Stock ${ticker} does not have an S&P ID.`);
         }
       }
     } else {
       // When no specific stock is requested, we fetch all stocks from the database which have a Standard & Poor’s ID.
-      [stocks] = await readAllStocks({
+      [stockList] = await readAllStocks({
         where: {
           spID: {
             not: null,
@@ -1407,7 +375,7 @@ export class FetchController {
       });
     }
 
-    if (stocks.length === 0) {
+    if (stockList.length === 0) {
       // If no stocks are left, we return a 204 No Content response.
       res.status(204).end();
       return;
@@ -1417,115 +385,18 @@ export class FetchController {
       res.status(202).end();
     }
 
-    const updatedStocks: Stock[] = [];
-    let successfulCount = 0;
-    let errorCount = 0;
-    let consecutiveErrorCount = 0;
-    const driver = await getDriver(true);
-    for await (const stock of stocks) {
-      if (
-        !req.query.noSkip &&
-        stock.spLastFetch &&
-        // We only fetch stocks that have not been fetched in the last 7 days.
-        new Date().getTime() - stock.spLastFetch.getTime() < 1000 * 60 * 60 * 24 * 7
-      ) {
-        logger.info(
-          PREFIX_SELENIUM +
-            `Stock ${stock.ticker}: Skipping S&P fetch because last fetch was ${formatDistance(
-              stock.spLastFetch.getTime(),
-              new Date().getTime(),
-              { addSuffix: true }
-            )}`
-        );
-        continue;
-      }
-      let spESGScore: number = req.query.clear ? null : undefined;
+    const stocks: FetcherWorkspace<Stock> = {
+      queued: [...stockList],
+      successful: [],
+      failed: [],
+    };
 
-      try {
-        await driver.get(`https://www.spglobal.com/esg/scores/results?cid=${stock.spID}`);
-        // Wait for the page to load for a maximum of 5 seconds.
-        await driver.wait(until.elementLocated(By.xpath(XPATH_SP_PANEL)), 5000);
+    await Promise.all([...Array(determineConcurrency(req))].map(() => spFetcher(req, stocks)));
 
-        const lockedContent = await driver.findElements(By.className("lock__content"));
-        if (
-          lockedContent.length > 0 &&
-          (await lockedContent[0].getText()).includes(
-            "This company's ESG Score and underlying data are available via our premium channels"
-          )
-        ) {
-          // If the content is available for premium subscribers only, we throw an error.
-          // Sadly, we are not a premium subscriber :(
-          throw new Error("This stock’s ESG Score is available for S&P Premium subscribers only.");
-        }
-
-        spESGScore = +(await (await driver.findElement(By.id("esg-score"))).getText());
-
-        // Update the stock in the database.
-        await updateStock(stock.ticker, {
-          spLastFetch: new Date(),
-          spESGScore,
-        });
-        updatedStocks.push(await readStock(stock.ticker));
-        successfulCount += 1;
-        consecutiveErrorCount = 0;
-      } catch (e) {
-        if (req.query.ticker) {
-          // If this request was for a single stock, we shut down the driver and throw an error.
-          await quitDriver(driver);
-          throw new APIError(
-            502,
-            `Stock ${stock.ticker}: Unable to fetch S&P ESG Score: ${String(e.message).split(/[\n:{]/)[0]}`
-          );
-        }
-        logger.warn(PREFIX_SELENIUM + chalk.yellowBright(`Stock ${stock.ticker}: Unable to fetch S&P ESG Score: ${e}`));
-        if (stock.spESGScore !== null) {
-          // If an S&P ESG Score is already stored in the database, but we cannot extract it from the page, we log this
-          // as an error and send a message.
-          logger.error(
-            PREFIX_SELENIUM +
-              chalk.redBright(
-                `Stock ${stock.ticker}: Extraction of S&P ESG Score failed unexpectedly. ` +
-                  `This incident will be reported.`
-              )
-          );
-          await signal.sendMessage(
-            SIGNAL_PREFIX_ERROR +
-              `Stock ${stock.ticker}: Unable to fetch S&P ESG Score: ${
-                String(e.message).split(/[\n:{]/)[0]
-              }\n${await takeScreenshot(driver, stock, "sp")}`,
-            "fetchError"
-          );
-          await new Promise((resolve) => setTimeout(resolve, 3000)); // Cool down for 3 seconds.
-          errorCount += 1;
-          consecutiveErrorCount += 1;
-        } else {
-          successfulCount += 1;
-          consecutiveErrorCount = 0;
-        }
-      }
-      if (consecutiveErrorCount >= 5) {
-        // If we have 5 consecutive errors, we stop fetching data, since something is probably wrong.
-        logger.error(
-          PREFIX_SELENIUM +
-            chalk.redBright(
-              `Aborting fetching information from S&P after ${consecutiveErrorCount} consecutive failures, ` +
-                `${successfulCount} successful fetches and ${errorCount} total failures. Will continue next time.`
-            )
-        );
-        await signal.sendMessage(
-          SIGNAL_PREFIX_ERROR +
-            `Aborting fetching information from S&P after ${consecutiveErrorCount} consecutive failures, ` +
-            `${successfulCount} successful fetches and ${errorCount} total failures. Will continue next time.`,
-          "fetchError"
-        );
-        break;
-      }
-    }
-    await quitDriver(driver);
-    if (updatedStocks.length === 0) {
+    if (stocks.successful.length === 0) {
       res.status(204).end();
     } else {
-      res.status(200).json(updatedStocks).end();
+      res.status(200).json(stocks.successful).end();
     }
   }
 
@@ -1542,21 +413,21 @@ export class FetchController {
     accessRights: GENERAL_ACCESS + WRITE_STOCKS_ACCESS,
   })
   async fetchSustainalyticsData(req: Request, res: Response) {
-    let stocks: Stock[];
+    let stockList: Stock[];
 
     if (req.query.ticker) {
       // A single stock is requested.
       const ticker = req.query.ticker;
       if (typeof ticker === "string") {
-        stocks = [await readStock(ticker)];
-        if (!stocks[0].sustainalyticsID) {
+        stockList = [await readStock(ticker)];
+        if (!stockList[0].sustainalyticsID) {
           // If the only stock to use does not have a Sustainalytics ID, we throw an error.
           throw new APIError(404, `Stock ${ticker} does not have a Sustainalytics ID.`);
         }
       }
     } else {
       // When no specific stock is requested, we fetch all stocks from the database which have a Sustainalytics ID.
-      [stocks] = await readAllStocks({
+      [stockList] = await readAllStocks({
         where: {
           sustainalyticsID: {
             not: null,
@@ -1565,7 +436,7 @@ export class FetchController {
       });
     }
 
-    if (stocks.length === 0) {
+    if (stockList.length === 0) {
       // If no stocks are left, we return a 204 No Content response.
       res.status(204).end();
       return;
@@ -1578,7 +449,6 @@ export class FetchController {
     const updatedStocks: Stock[] = [];
     let successfulCount = 0;
     let errorCount = 0;
-    let consecutiveErrorCount = 0;
     let sustainalyticsXMLResource: Resource;
     try {
       try {
@@ -1631,7 +501,7 @@ export class FetchController {
 
     const sustainalyticsXMLLines = sustainalyticsXMLResource.content.split("\n");
 
-    for await (const stock of stocks) {
+    for await (const stock of stockList) {
       let sustainalyticsESGRisk: number = req.query.clear ? null : undefined;
 
       try {
@@ -1665,7 +535,6 @@ export class FetchController {
         });
         updatedStocks.push(await readStock(stock.ticker));
         successfulCount += 1;
-        consecutiveErrorCount = 0;
       } catch (e) {
         if (req.query.ticker) {
           // If this request was for a single stock, we shut down the driver and throw an error.
@@ -1695,27 +564,23 @@ export class FetchController {
             "fetchError"
           );
           errorCount += 1;
-          consecutiveErrorCount += 1;
         } else {
           successfulCount += 1;
-          consecutiveErrorCount = 0;
         }
       }
-      if (consecutiveErrorCount >= 10) {
-        // If we have 10 consecutive errors, we stop fetching data, since something is probably wrong.
+      if (errorCount >= 10) {
+        // If we have 10 errors, we stop fetching data, since something is probably wrong.
         logger.error(
           PREFIX_SELENIUM +
             chalk.redBright(
-              `Aborting extracting information from Sustainalytics after ${consecutiveErrorCount} ` +
-                `consecutive failures, ${successfulCount} successful fetches and ${errorCount} total failures. ` +
-                `Will continue next time.`
+              `Aborting extracting information from Sustainalytics after ${successfulCount} successful fetches ` +
+                `and ${errorCount} failures. Will continue next time.`
             )
         );
         await signal.sendMessage(
           SIGNAL_PREFIX_ERROR +
-            `Aborting extracting information from Sustainalytics after ${consecutiveErrorCount} ` +
-            `consecutive failures, ${successfulCount} successful fetches and ${errorCount} total failures. ` +
-            `Will continue next time.`,
+            `Aborting extracting information from Sustainalytics after ${successfulCount} successful fetches ` +
+            `and ${errorCount} failures. Will continue next time.`,
           "fetchError"
         );
         break;
