+++ packages/rating-tracker-backend/src/openapi/paths/fetch/marketScreenerEndpoint.ts
@@ -23,6 +23,7 @@ const post: OpenAPIV3.OperationObject = {
     fetch.detach,
     fetch.noSkip,
     fetch.clear,
+    fetch.concurrency,
   ],
   responses: {
     "200": okStockList,
