+++ packages/rating-tracker-backend/src/utils/webdriver.ts
@@ -1,0 +1,92 @@
+// This class is not tested because it is not possible to use it without a running Selenium WebDriver.
+/* istanbul ignore file -- @preserve */
+import { Builder, Capabilities, WebDriver } from "selenium-webdriver";
+import * as chrome from "selenium-webdriver/chrome.js";
+import APIError from "./apiError.js";
+import logger, { PREFIX_SELENIUM } from "./logger.js";
+import chalk from "chalk";
+import { Stock } from "rating-tracker-commons";
+import { createResource } from "../redis/repositories/resourceRepository.js";
+
+/**
+ * A page load strategy to use by the WebDriver.
+ */
+type PageLoadStrategy = "normal" | "eager" | "none";
+
+/**
+ * Creates and returns a new WebDriver instance.
+ *
+ * @param {boolean} headless whether to run the browser in headless mode
+ * @param {PageLoadStrategy} pageLoadStrategy whether to run the browser in headless mode
+ * @returns {Promise<WebDriver>} a Promise that resolves to a WebDriver instance
+ * @throws an {@link APIError} if the WebDriver cannot be created
+ */
+export const getDriver = async (headless?: boolean, pageLoadStrategy?: PageLoadStrategy): Promise<WebDriver> => {
+  const url = process.env.SELENIUM_URL;
+  const options = new chrome.Options()
+    .addArguments("window-size=1080x3840") // convenient for screenshots
+    .addArguments("--blink-settings=imagesEnabled=false"); // Do not load images
+  headless && options.headless(); // In headless mode, the browser window is not shown.
+
+  return await new Builder()
+    .usingServer(url)
+    .withCapabilities(
+      new Capabilities()
+        // Use Chrome as the browser.
+        .setBrowserName("chrome")
+        // Do not wait for all resources to load. This speeds up the page load.
+        .setPageLoadStrategy(pageLoadStrategy ?? "none")
+    )
+    .setChromeOptions(options)
+    .build()
+    .then((driver) => driver)
+    .catch((e) => {
+      throw new APIError(502, `Unable to connect to Selenium WebDriver: ${e.message}`);
+    });
+};
+
+/**
+ * Shuts down the given WebDriver instance gracefully, deallocating all associated resources.
+ *
+ * @param {WebDriver} driver the WebDriver instance to shut down
+ * @returns {Promise<void>} a Promise that resolves when the WebDriver has been shut down
+ * @throws an {@link APIError} if the WebDriver cannot be shut down gracefully
+ */
+export const quitDriver = async (driver: WebDriver): Promise<void> => {
+  try {
+    await driver.quit();
+  } catch (e) {
+    logger.warn(PREFIX_SELENIUM + chalk.yellowBright(`Unable to shut down Selenium WebDriver gracefully: ${e}`));
+  }
+};
+
+/**
+ * Creates a screenshot of the current page and stores it in Redis.
+ *
+ * @param {WebDriver} driver the WebDriver instance in use
+ * @param {Stock} stock the affected stock
+ * @param {string} dataProvider the name of the data provider
+ * @returns {Promise<string>} a Promise that resolves to a string holding a general informational message and a URL to
+ * the screenshot
+ */
+export const takeScreenshot = async (driver: WebDriver, stock: Stock, dataProvider: string): Promise<string> => {
+  const screenshotID = `error-${dataProvider}-${stock.ticker}-${new Date().getTime().toString()}.png`;
+  try {
+    const screenshot = await driver.takeScreenshot();
+    // deepcode ignore Ssrf: This is a custom function named `fetch()`, which does not perform a request
+    await createResource(
+      {
+        url: screenshotID,
+        fetchDate: new Date(),
+        content: screenshot, // base64-encoded PNG image
+      },
+      60 * 60 * 24 // We only store the screenshot for 24 hours.
+    );
+    return `For additional information, see https://${process.env.SUBDOMAIN ? process.env.SUBDOMAIN + "." : ""}${
+      process.env.DOMAIN
+    }/api/resource/${screenshotID}.`;
+  } catch (e) {
+    logger.warn(PREFIX_SELENIUM + chalk.yellowBright(`Unable to take screenshot “${screenshotID}”: ${e}`));
+    return "";
+  }
+};
